refactor(store-core): extract session-status handlers into handlers/session-status.ts (audit P2-3)
Eleventh slice of the P2-3 split — first sub-family of the session-lifecycle
core. Move the session-status / client-focus handlers (session_context,
session_timeout, session_restore_failed, session_warning, session_switched,
client_focus_changed) and their payload interfaces into a new
handlers/session-status.ts.

Pure move, no logic change: the moved block is byte-identical to the
origin/main block (proven by diff). Imports ChatMessage (../types),
nextMessageId (../utils), parseStringField + resolveSessionId + SessionPatch
(./_shared); ClientFocusChanged is local. All index.ts imports stay used.
The barrel re-exports via `export * from './session-status'`, so the public
surface is unchanged.

handlers/index.ts: 3371 -> 3137 lines.
